server/requirements.txt
fastapi[standard]
Pillow
rembg
